+++ crest/Assets/Crest/Crest/Shaders/Underwater/UnderwaterCurtain.shader
@@ -184,7 +184,7 @@ Shader "Crest/Underwater Curtain"
 
 				const float pixelZ = LinearEyeDepth(input.positionCS.z);
 				const half3 screenPos = input.foam_screenPos.yzw;
-				const half2 uvDepth = screenPos.xy / screenPos.z;
+				const float2 uvDepth = screenPos.xy / screenPos.z;
 				const float sceneZ01 = SAMPLE_TEXTURE2D_X(_CameraDepthTexture, sampler_CameraDepthTexture, uvDepth).x;
 				const float sceneZ = LinearEyeDepth(sceneZ01);
